+++ CHANGES.rst
@@ -1,6 +1,7 @@
 0.2 (Unreleased)
 ================
 
+- Add configurable region/language preferences [#7]
 - Ensure package name is valid for JDownloader [#6]
 
 0.1 (2024-10-19)